+++ R/utils_set_attr.R
@@ -1,3 +1,4 @@
+#' @importFrom insight link_inverse link_function
 .set_attributes_and_class <- function(data, model, t.title, x.title, y.title, l.title, legend.labels, x.axis.labels, model_info, constant.values = NULL, terms = NULL, original_terms = NULL, at_list = NULL, n.trials = NULL, prediction.interval = NULL, condition = NULL, ci.lvl = .95) {
   # check correct labels
   if (!is.null(x.axis.labels) && length(x.axis.labels) != length(stats::na.omit(unique(data$x))))
@@ -27,6 +28,8 @@
   attr(data, "family") <- model_info$family
   attr(data, "link") <- model_info$link_function
   attr(data, "logistic") <- ifelse(model_info$is_binomial || model_info$is_ordinal || model_info$is_multinomial, "1", "0")
+  attr(data, "link_inverse") <- insight::link_inverse(model)
+  attr(data, "link_function") <- insight::link_function(model)
   attr(data, "is.trial") <- ifelse(model_info$is_trial && inherits(model, "brmsfit"), "1", "0")
   attr(data, "n.trials") <- n.trials